requirements.txt
apache-airflow==2.1.1
apache-airflow-providers-postgres==2.3.0
slack==0.0.2
slackclient==2.9.3
slack-sdk==3.11.2
pandas==1.3.1      
